// src/API.ts
export type Todo = {
  __typename: 'Todo';
  id: string;
  title: string;
  message: string;
  isDone: boolean;
  dueDate?: string | null;
  createdAt: string;
  type: string;
  updatedAt: string;
  owner?: string | null;
};